requirements.txt
Flask>=2.1.0
requests>=2.28.0
pybaseball>=2.2.0
google-cloud-storage>=2.7.0
google-cloud-bigquery>=3.12.0
pandas>=1.5.0
joblib>=1.2.0
xgboost>=1.7.0
scikit-learn>=1.1.0
optuna>=3.0.0
